requirements.txt
numpy>=1.21.0
tensorflow>=2.10.0
matplotlib>=3.5.0
pillow>=9.0.0
